Maintain | xv6: update .clang_complete and fix warning

+++ xv6-public/syscall.c
@@ -167,7 +167,7 @@ syscall(void)
   struct proc *curproc = myproc();
 
   num = curproc->tf->eax;
-  if(num > 0 && num < NELEM(syscalls) && syscalls[num]) {
+  if(num > 0 && num < (int)NELEM(syscalls) && syscalls[num]) {
     curproc->tf->eax = syscalls[num]();
     /* cprintf("%s -> %d\n", syscalls_name[num], curproc->tf->eax); */
   } else {
